Refactor library into package
