+++ examples/py/PY3-filter_interaction.py
@@ -1,12 +1,13 @@
-import sys, os
-sys.path.append(os.path.abspath('../../src'))
+import os
+from pathlib import Path
+from dotenv import load_dotenv
+load_dotenv(Path.cwd().parent / ".env")
+env = os.environ
 
 from bokeh.models import WheelZoomTool
-from visualization.st_visualizer import st_visualizer
-import express.st_vizexpress as viz_express
-from dotenv import load_dotenv
-load_dotenv("..\.env")
-env = os.environ
+from st_visions.visualization.st_visualizer import st_visualizer
+import st_visions.express.st_vizexpress as viz_express
+
 
 # TO EXECUTE SCRIPT USE (ON REMOTE SERVER)
 # python -m bokeh serve --show examples/py/PY3-filter_interaction.py --allow-websocket-origin=<NODE_IP_ADDRESS>:<BOKEH_PORT>
