+++ st_visions/visualization/filter_callbacks.py
